Rename the project data refinery.
