+++ README.md
@@ -1,2 +1,3 @@
-# bioinformatics_mill
-A project to download, process, aggregate, and serve bioinformatic data supported by Green Lab.
+# data_refinery
+A project to download, process, aggregate, and serve bioinformatic data
+supported by Green Lab.
