+++ data_models/data_refinery_models/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for bioinformatics_mill_workers project.
+Django settings for data_refinery_models project.
 
 Generated by 'django-admin startproject' using Django 1.10.6.
 
@@ -47,8 +47,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'bioinformatics_mill_models',
-    'downloaders',
+    'data_refinery_models',
 ]
 
 MIDDLEWARE = [
@@ -61,7 +60,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'bioinformatics_mill_workers.urls'
+ROOT_URLCONF = 'data_refinery_models.urls'
 
 TEMPLATES = [
     {
@@ -79,11 +78,12 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'bioinformatics_mill_workers.wsgi.application'
+WSGI_APPLICATION = 'data_refinery_models.wsgi.application'
 
 
 # Database
 # https://docs.djangoproject.com/en/1.10/ref/settings/#databases
+
 
 DATABASES = {
     'default': {
@@ -96,9 +96,6 @@ DATABASES = {
     }
 }
 
-
-# Celery
-CELERY_BROKER_URL = get_env_variable('CELERY_BROKER_URL')
 
 # Password validation
 # https://docs.djangoproject.com/en/1.10/ref/settings/#auth-password-validators
