+++ data_models/data_refinery_models/urls.py
@@ -1,4 +1,4 @@
-"""bioinformatics_mill_models URL Configuration
+"""data_refinery_models URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.10/topics/http/urls/
