+++ data_models/data_refinery_models/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for bioinformatics_mill_workers project.
+WSGI config for data_refinery_models project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,7 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "bioinformatics_mill_workers.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE",
+                      "data_refinery_models.settings")
 
 application = get_wsgi_application()
