+++ data_models/install.sh
@@ -7,11 +7,11 @@
 # (i.e. 'sudo POSTGRES_VERSION=9.4 -u postgres install.sh')
 if [ -z $POSTGRES_VERSION ]; then POSTGRES_VERSION=9.5; fi
 
-createdb bioinformatics_mill
+createdb data_refinery
 
-psql bioinformatics_mill -c "CREATE ROLE bioinformatics_mill_user WITH LOGIN PASSWORD 'bioinformatics_password';"
-psql bioinformatics_mill -c 'GRANT ALL PRIVILEGES ON DATABASE bioinformatics_mill TO bioinformatics_mill_user;'
-psql bioinformatics_mill -c 'ALTER USER bioinformatics_mill_user CREATEDB;'
+psql data_refinery -c "CREATE ROLE data_refinery_user WITH LOGIN PASSWORD 'data_refinery_password';"
+psql data_refinery -c 'GRANT ALL PRIVILEGES ON DATABASE data_refinery TO data_refinery_user;'
+psql data_refinery -c 'ALTER USER data_refinery_user CREATEDB;'
 
 iptables -A INPUT -s 172.17.0.0/16 -j ACCEPT
