+++ data_models/make_migrations.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+env $(cat environments/dev | xargs) ./manage.py makemigrations data_refinery_models
+env $(cat environments/dev | xargs) ./manage.py migrate
