+++ data_models/manage.py
@@ -3,7 +3,8 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "bioinformatics_mill_models.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE",
+                          "data_refinery_models.settings")
     try:
         from django.core.management import execute_from_command_line
     except ImportError:
