+++ data_models/setup.py
@@ -5,12 +5,12 @@ from setuptools import find_packages, setup
 os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
 
 setup(
-    name='bioinformatics-mill-models',
+    name='data-refinery-models',
     version='0.1',
     packages=find_packages(),
     include_package_data=True,
     license='BSD License',
-    description='An abstracted database model for the Bioinformatics Mill project',
+    description='An abstracted database model for the Data Refinery project',
     url='https://www.greenelab.com',
     author='Kurt Wheeler',
     author_email='kurt.wheeler91@gmail.com',
