+++ workers/data_refinery_workers/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for bioinformatics_mill_models project.
+Django settings for data_refinery_workers project.
 
 Generated by 'django-admin startproject' using Django 1.10.6.
 
@@ -47,7 +47,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'bioinformatics_mill_models',
+    'data_refinery_models',
+    'downloaders',
 ]
 
 MIDDLEWARE = [
@@ -60,7 +61,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'bioinformatics_mill_models.urls'
+ROOT_URLCONF = 'data_refinery_workers.urls'
 
 TEMPLATES = [
     {
@@ -78,12 +79,11 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'bioinformatics_mill_models.wsgi.application'
+WSGI_APPLICATION = 'data_refinery_workers.wsgi.application'
 
 
 # Database
 # https://docs.djangoproject.com/en/1.10/ref/settings/#databases
-
 
 DATABASES = {
     'default': {
@@ -96,6 +96,9 @@ DATABASES = {
     }
 }
 
+
+# Celery
+CELERY_BROKER_URL = get_env_variable('CELERY_BROKER_URL')
 
 # Password validation
 # https://docs.djangoproject.com/en/1.10/ref/settings/#auth-password-validators
