+++ workers/data_refinery_workers/urls.py
@@ -1,4 +1,4 @@
-"""bioinformatics_mill_workers URL Configuration
+"""data_refinery_workers URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.10/topics/http/urls/
