+++ workers/environments/test
@@ -1,9 +1,9 @@
 DJANGO_SECRET_KEY=()vc0@18d&m!@d#g&jy450pfp%%otf66f8t9hdjy)(q*n(nf+v
 DJANGO_DEBUG=True
 
-DATABASE_NAME=bioinformatics_mill
-DATABASE_USER=bioinformatics_mill_user
-DATABASE_PASSWORD=bioinformatics_password
+DATABASE_NAME=data_refinery
+DATABASE_USER=data_refinery_user
+DATABASE_PASSWORD=data_refinery_password
 DATABASE_HOST=database
 DATABASE_PORT=5432
