+++ workers/requirements.in
@@ -2,4 +2,3 @@ django
 celery
 psycopg2
 requests
-bioinformatics-mill-models
